+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# version
+## Someday
+
+ - fix "database is locked"
+
 # 1.2.0
 ## 2025-02-23